+++ emucoo-dto/src/main/java/com/emucoo/dto/modules/report/ChancePointVo.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Data
 public class ChancePointVo {
     private Integer chancePointFrequency;
-    private Integer chancePointID;
+    private Long chancePointID;
     private String chancePointTitle;
     private String checkItemContent;
     private List<ReportWorkVo> workArr;
